feat: move serialization code
