+++ python/beeai_framework/serializer/serializable.py
