+++ python/tests/serializer/test_clone.py
@@ -12,18 +12,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-from beeai_framework.memory.base_memory import BaseMemory
-from beeai_framework.memory.errors import ResourceError, ResourceFatalError
-from beeai_framework.memory.readonly_memory import ReadOnlyMemory
-from beeai_framework.memory.token_memory import TokenMemory
-from beeai_framework.memory.unconstrained_memory import UnconstrainedMemory
-
-__all__ = [
-    "BaseMemory",
-    "ReadOnlyMemory",
-    "ResourceError",
-    "ResourceFatalError",
-    "TokenMemory",
-    "UnconstrainedMemory",
-]
